+++ clients/sns-client/src/main/java/org/snsclient/twitter/constants/ApiUrls.java
@@ -4,4 +4,6 @@ public class ApiUrls {
 
 	public static final String TWITTER_MEDIA_UPLOAD_URL = "https://api.x.com/2/media/upload";
 	public static final String TWITTER_GET_ME_URL = "https://api.x.com/2/users/me";
+	public static final String TWITTER_GET_TOKEN_URL = "https://api.x.com/2/oauth2/token";
+	public static final String TWITTER_POST_TWEET_URL = "https://api.x.com/2/tweets";
 }
